Braille/backend/requirement.txt
flask
flask-cors
transformers
torch
torchvision
pillow
